cmd/compile: assign and preserve statement boundaries.
A pass run after ssa building (before any other optimization)
identifies the "first" ssa node for each statement.
Other "noise" nodes are tagged as being never appropriate
for a statement boundary (e.g., VarKill, VarDef, Phi).
Rewrite and deadcode are modified to move the statement
boundaries forward whenever possible if a boundary-tagged
ssa value is removed; never-boundary nodes are ignored in
this search (some operations involving constants are also
tagged as never-boundary and also ignored because they are
likely to be moved or removed during optimization).

Code generation treats all nodes except those explicitly
marked as statement boundaries as "not statement" nodes.

Line number html conversion was modified to make statement
boundary nodes a bit more obvious by prepending a "+".

The code in fuse.go that glued together the value slices
of two blocks produced a result that depended on the
former capacities (not lengths) of the two slices.  This
causes differences in the 386 bootstrap, and also can
sometimes put values into an order that does a worse job
of preserving statement boundaries when values are removed.

This CL worsens the optimized-debugging experience with Delve;
we need to work with the delve team to get so that they can
use the is_stmt marks that we're emitting now.

Portions of two delve tests that had caught problems were
incorporated into ssa/debug_test.go.  There are some
opportunities to do better with optimized code, but the
nest-ing is not lying or overly jumpy.

Over 4 CLs, compilebench geomean measured
binary size increase of 4% and compile time
(both user and wall) increase of 1.5%.
Most of this (3.2%, 0.9%) comes in the last CL.
Most of the user-time increase is in Flate.

Trybot failures for linux-amd64 did not reproduce on
linux-amd64 despite 10 or more non-cached repeat tests.

Change-Id: I359c94e06843f1eb41f9da437bd614885aa9644a