encoding/xml: fix unexpected behavior of encoder.Indent("", "")
MarshalIndent and Marshal share the code. When prefix and indent are
empty, the behaviour is like Marshal when it should have a minimal
indent, i.e. new line.

A boolean is added to the local printer struct which defaults to false.
It is set to true only when MarshalIndent is used and prefix and indent
are empty.

Comments have now a leading space.

The type structure in the _test.go is kept in line with the rest of the
file as the structs are used for one test each time.

Fixes #13185

Change-Id: I2835b88ddef41ee281f893119cbad010192f7dc7