Update patch set 1
Patch Set 1:

> Patch Set 1: -Run-TryBot
> 
> So what do you want me to do about this review? I do now know much about gdb. And I even cannot run the tests.
> 
> Alex

The main reason for this CL is to ensure that tests report the actual error with gdb 7.9 and above. (This part can be tested with TDM-GCC http://tdm-gcc.tdragon.net/)

As for possible resolutions for this CL:

1. Continue with this CL and make the actual error visible (the current CL).
2. Continue with this CL and make the actual error visible, but also skip gdb tests for the problematic gdb-s (not sure how to detect this or whether it should be done)
3. Find (or compile) gdb 7.8 or 7.7  with python support and use that in testing and build environments, then submit this CL with adjustments.
4. Find (or compile) newer gdb with python support and fix free_pendings, then submit this CL with other fixes.
5. Maybe, figure out why exactly newer gdb is crashing, adjust go compiler to ensure it doesn't trigger it.

But, as I see, the fixes to the Tests in this CL need to be done in all of these scenarios, therefore I thought this CL might be a good small step.

Patch-set: 1
Reviewer: Gerrit User 6785 <6785@62eb7196-b449-3ce5-99f1-c037f21e1705>