Update patch set 17
Patch Set 17:

(3 comments)

This is a lot of new code for something that's slower.

Like I said before, this probably needs a cut-off to use the old-vs-new algorithm if the data is small.

Also, this would need an update of zfuncversion.go.

Can you mark this as DO NOT REVIEW until it's ready?

Patch-set: 17