go/build, go/types, cmd/dist: add js/wasm architecture
This is the first commit of a series that will add WebAssembly
as an architecture target. The design document can be found at
https://docs.google.com/document/d/131vjr4DH6JFnb-blm_uRdaC0_Nv3OUwjEY5qVCxCup4.

The GOARCH name "wasm" is the official abbreviation of WebAssembly.
The GOOS name "js" got chosen because WebAssembly code accesses the
environment via the WebAssembly runtime. Initially this runtime will
be web browsers and Node.js, which both use JavaScript to embed
WebAssembly. Other GOOS values may be possible later, see
https://github.com/WebAssembly/design/blob/master/NonWeb.md for more
information.

Updates #18892

Change-Id: Ia25b4fa26bba8029c25923f48ad009fd3681933a