Update patch set 4
Patch Set 4:

Thanks, Daniel.

I like to keep this on hold for the moment (that is to say I am not ignoring this).

There are some other changes I need to make to go/types with respect to handling of cycles and how incorrect types should be dealt with. Maybe this falls out of that, or maybe this is the right addition to that work. Either way, this is not urgent, and we have more time during the next freeze do deal with small bug fixes.

Patch-set: 4
Reviewer: Gerrit User 5210 <5210@62eb7196-b449-3ce5-99f1-c037f21e1705>