cmd/go: fix "go build -a" tests
CL 73212 and CL 76015 changed the go command to base rebuilding
decisions on file contents rather than time, and removed
"runtime" as a dependency of every package.
This broke TestGoBuildDashA* tests.

This CL fixed those tests by using the "time" package as a subject
because "runtime" is not a dependency of "math" anymore.
This test was changed to run in parallel.

Also, environment variable TESTGO_IS_GO_RELEASE is not used anywhere,
thus testing of "go build -a" was reduced to a general case.

Updates #24436

Change-Id: I17fa8c22be1b35d70ec051b69a150cc1625c1285