cmd/go: fix build -a tests
CL 73212 and CL 76015 made the go command to base all its rebuilding
decisions on the content of the files being processed and removed
"runtime" from the default list of dependencies of every package.
This made some tests to be broken.

This CL changed tests to use the "time" package instead of "math".

Environment variable TESTGO_IS_GO_RELEASE is not used anymore, thus
testing "build -a" can be reduced to a general case.

Updates #24436

Change-Id: I17fa8c22be1b35d70ec051b69a150cc1625c1285