Update patch set 4
Patch Set 4:

> What is the status of this change now?

This CL fixes code that is removed now. So this CL can be abandoned.

Jeff Johnson was going to try again to fix #20073, so we might find this code useful for that. But we could un-abandon this code then.

Alex

Patch-set: 4
Reviewer: Gerrit User 5070 <5070@62eb7196-b449-3ce5-99f1-c037f21e1705>