Update patch set 13
Patch Set 13:

> Patch Set 13:
> 
> > Patch Set 13:
> > 
> > > Patch Set 13:
> > > 
> > > > Patch Set 13:
> > > > 
> > > > Same failure. Even more evidence that it's real. I see no evidence that it's memory related.
> > > 
> > > Thank you. I might need to setup a real 386 machine to check what's wrong.
> > 
> > We don't have real 386 machines. We just set GOHOSTARCH=386 GOARCH=386: https://github.com/golang/build/blob/3d62d0f/dashboard/builders.go#L913
> 
> Thank you. I would do more checks.

I have a theory as to what is going wrong.
a MOVSD doesn't have to have an aligned memory source, but an ADDSD does.  float64s are aligned correctly on amd64 but only to 4 bytes on 386.

Try removing the *SD rules and see if that helps.

Patch-set: 13
Reviewer: Gerrit User 5200 <5200@62eb7196-b449-3ce5-99f1-c037f21e1705>