Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> (1 comment)
> 
> Just a note about unrelated changes.

I'm not yet mastering the history of commits in multiple branches. So I threw away the whole thing (which was small) and re-submitted a new patch.

I'm still interested to know how to submit an optimization. I've seen a few things in encoding/xml worth simplifying.

Regards,

Patch-set: 1