cmd/compile: add initial README
As a follow-up to the first README for cmd/compile/internal/ssa.

Since this is the parent package for all the compiler packages, this
README serves as an overview of the compiler and its packages. As more
READMEs are added for specific parts with more detail, such as ssa's,
they can be linked from this one.

Thanks to Iskander Sharipov, Josh Bleecher Snyder, and Matthew Dempsky
for helping with all the details in this document.

Change-Id: I820a535e25dce86ccc667ce1c6e92b75fc32f3af