cmd/compile: in prove, complete support for OpIsInBounds/OpIsSliceInBounds
The logic in updateBranchRestrictions didn't allow to correctly
model OpIs(Slice)Bound for signed domain, and it was also partly
implemented within updateRestrictions.

Thanks to the previous changes, it is now possible to handle
the negative conditions correctly, so that we can learn
both signed/LT + unsigned/LT on the positive side, and
signed/GE + unsigned/GE on the negative side (but only if
the index can be proved to be non-negative).

Change-Id: I9858080dc03b16f85993a55983dbc4b00f8491b0