Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Could this be refactored to use findIndVar()? I've sent you a patch yesterday to demonstrate the approach.
> 
> If we want to get rid of loopbce, we need to use findIndVar() anyway and it looks very useful to have an utility to detect all induction variables decouple from prove itself. Pushing findIndVar information into prove is quite easy then.

I'll try to take a look at your CLs in the next couple days.

For this CL, it would be difficult to factor the test into findIndVar because it depends on the factsTable to test the min < max condition (which is important for the OFORUNTIL form because the loop isn't dominated by a single check that ind < max).

findIndVar and the test introduced by this CL are looking for fairly different patterns. Obviously it would be nice to unify them, but I'm also not too concerned about it.

Patch-set: 1
Reviewer: Gerrit User 5167 <5167@62eb7196-b449-3ce5-99f1-c037f21e1705>