Update patch set 5
Patch Set 5:

> I think you're fine to submit. It could be worth looking into why Unicode went up, but I'm not majorly concerned.

I have narrowed it down to two lines - reverting this bit removes the extra allocs:

diff --git a/src/cmd/compile/internal/gc/sinit.go b/src/cmd/compile/internal/gc/sinit.go
index b63ac23ae3..0ac75af148 100644
--- a/src/cmd/compile/internal/gc/sinit.go
+++ b/src/cmd/compile/internal/gc/sinit.go
@@ -451,7 +451,8 @@ func staticassign(l *Node, r *Node, out *[]*Node) bool {
 				gdata(n, e.Expr, int(n.Type.Width))
 			} else {
 				setlineno(e.Expr)
-				a := n.copy()
+				a := nod(OXXX, nil, nil)
+				*a = *n
 				a.Orig = a // completely separate copy
 				if !staticassign(a, e.Expr, out) {
 					*out = append(*out, nod(OAS, a, e.Expr))


Playing with -gcflags=-m, it seems like before gc saw this case's n copy as non-escaping, but now it fails to do so. I presume this is because the code is a bit more complex now, with the addition of method calls.

As far as I can tell, the old and new code are otherwise equivalent, and as you say this is a minor bump that doesn't seem to affect performance. I'm going to submit for now.

If you think this should be a bug, or if it should be fixed, please let me know. But my assupmtion is that we can just chuck this at "improve escape analysis". Maybe I should post this as an example to improve somewhere on the issue tracker.

Patch-set: 5
Reviewer: Gerrit User 13550 <13550@62eb7196-b449-3ce5-99f1-c037f21e1705>