encoding/xml: fix normalization of attributes values
The attribute value was read as text. The existing attribute reader logic is fixed as an attribute may have a namespace or only a prefix. Other possibilities have been removed.

To keep the behaviour of raw token which allows many faults in attributes list, error handling is heavily using the Strict parameter of the decoder. Specific tests have been added including list of attributes.

To keep the behaviour or unmarshal, escaped characters handling has been added but it is not symmetrical to Marshal for quotes but follows XML specification. Testing has been extended.

Fixes #20614

Change-Id: Ic4014c4cd5c44f6460de9e4fdbb5a34596575a96