cmd/link: do not pass -no-pie to host linker when -linkshared is passed
As the comment above the code I'm changing, when building with -buildmode=exe,
the default compiler flags produce code incompatible with PIE. But when
-linkshared is passed, the default compiler flags are not used so this does not
apply. And now I've found a system (linux/arm64 with glibc 2.27) where this
combination of flags causes a problem, albeit for reasons I don't really
understand, so stop passing -no-pie when -linkshared is passed.

Change-Id: I412ec7941dc0cb89e6d1b171fc29288aadcb9f20