Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+2
> 
> Assuming it's tested LGTM. But is "quirk" too judgemental? Is it quirky or just different from Linux? Maybe "alt"?

It's tested on the solaris-amd64-oraclerel and solaris-amd64-smartosbuildlet gomote builders and on freebsd/dragonfly locally.

I was going for "quirk" because the comments in the file mentioned it as such :-) But yes, "alt" would be less judgmental.

Patch-set: 1
Reviewer: Gerrit User 19560 <19560@62eb7196-b449-3ce5-99f1-c037f21e1705>