cmd/go: fix go get VCS tests that match go-import's
The existing test TestMatchGoImport asserts whether there should or
should not be an error. But in the case that that assertion passes and
there is an error (i.e. it's expected and we have one) we fail to assert
that the actual error matches the expected error in terms of content.

As a result there were a number of incorrect cases in the existing table
of tests for TestMatchGoImport that potentially obscured the scenario
that arose in https://github.com/golang/go/issues/24751

Change-Id: I1ea77f8f191dde8405d289de2386f26b9d074c14