Update patch set 20
Patch Set 20:

> Patch Set 20:
> 
> > Patch Set 20:
> > 
> > > Patch Set 20:
> > > 
> > > > Patch Set 20:
> > > > 
> > > > This may have broken the 387 builder; it has been failing consistently
> > > > since this change was merged: https://build.golang.org
> > > 
> > > You are right, I've sent https://go-review.googlesource.com/c/go/+/105035 as a fix
> > 
> > Why I did not find this failure on my localhost test? Due to the default GO386 is "sse2"? So I should do both "GOARCH=386 GOHOSTARCH=386 GO386=387" and ""GOARCH=386 GOHOSTARCH=386 GO386=sse2"next time?
> > 
> > Thank you for reviewing and fixing bugs in my CL. Really appreciate you.
> 
> Yes, you need to run both. I'm not sure why we don't have a trybot for GO386=387.
> Thanks for you contributions!

I'm still a bit worried about doing this for float64, because of alignment.
Bat at the same time, I can't get my machine to fail with an unaligned access.
I'll add a test to make sure we're still ok.

Ilya, are 386 and 387 ok with the combined load/add of float64s (into sse and mmx registers, respectively) if the load is not aligned?
I know some SSE instructions fail if the source is unaligned (PAND is the one I ran into long ago), not sure about the restrictions for float ops in 32-bit mode.

Patch-set: 20