Update patch set 7
Patch Set 7:

(1 comment)

Why don't we do this in two steps?  One CL to initialize internal/cpu as part of schedinit, and a separate CL to actually use that information.

Initializing internal/cpu as part of schedinit needs clearer docs.  We need to say exactly when initFeatures is called.  This would be best done by having a single initFeatures function that calls the GOARCH-specific one.  Then there is a place for the docs.

The name in the runtime package should be the name in the internal/cpu package; don't give it one name in one place and a different name in the other.  The function name should be something like internal_cpu_initFeatures.  E.g., the names used in sync/runtime.go.

Patch-set: 7
Reviewer: Gerrit User 5206 <5206@62eb7196-b449-3ce5-99f1-c037f21e1705>