cmd/compile: don't use packages of keys of composite literals prematurely
At parse time we don't know what a key k in a composite literal T{k: v}
refers to; it may be a field name, or an expression value. If k is a
name (an identifier), it may even be dot-imported. In the latter case,
the existing code marked the respective package as 'used', which let to
the omission of a "imported but not used" error.

The new code doesn't mark packages of keys that are dot-imported names.

Fixes #6428.

Change-Id: I32e4bf698907d329e048d6df8a3a1e38a40e5bb0