Update patch set 31
Patch Set 31:

(9 comments)

I ran a coverage report on this CL and the new
func stable(data Interface, n int) has only 77%
line coverage, e.g. its last two else blocks are
not executed during go test at all.
Fixing this will be complicated as you will have to
handcraft lots of testcases to handle all the control
flow. Formally it would be simpler to neglect the movement
imitation buffer and implement the actual algorithm from
Kim and Kutzner but this probably would not be much faster
(if at all for realistic block sizes). A MIB could be
introduced as a second step, speeding up the implementation.

But as I asked already: Are we sure _this_ algorithm from
Kim and Kutzner is the best for merging blocks of _equal_
sizes using a MIB? I could imagine a much simpler and faster
algorithm.

Patch-set: 31