Update patch set 32
Patch Set 32:

(7 comments)

> Patch Set 31:
> 
> (9 comments)
> 
> I ran a coverage report on this CL and the new
> func stable(data Interface, n int) has only 77%
> line coverage, e.g. its last two else blocks are
> not executed during go test at all.

I assume you just ran "go test" without modifying the "sizes :=" line in testBentleyMcIlroy? Much more sizes than the few present need to be added for exhaustive testing. That said, I did not try testing with coverage analysis, I will try that (with more sizes in the sizes slice) later.

> ... Formally it would be simpler to neglect the movement
> imitation buffer and implement the actual algorithm from
> Kim and Kutzner but this probably would not be much faster
> (if at all for realistic block sizes).

I do not understand this part of your message. In my code the MIB is used as in the paper, for block rearrangement. To what difference between the paper and my code are you alluding? 

> A MIB could be
> introduced as a second step, speeding up the implementation.

Second step of what?

> But as I asked already: Are we sure _this_ algorithm from
> Kim and Kutzner is the best for merging blocks of _equal_
> sizes using a MIB? I could imagine a much simpler and faster
> algorithm.

It is the best I know of, if you know a better one describe it or reference a paper; and presumably implement it.

Patch-set: 32
Reviewer: Gerrit User 15048 <15048@62eb7196-b449-3ce5-99f1-c037f21e1705>