Update patch set 2
Patch Set 2:

(18 comments)

The doc is basically rewritten in this patch set, so please give it another read. Hopefully the overall structure and splitting of compiler parts is more satisfactory now.

I've realised that even Josh and Matthew aren't clear on what backend and frontend mean exactly (and let's not add middle-end in the mix), so I've excluded that terminology from the meat of the doc. Still mentioned so that the reader has a rough idea of their meaning, though.

There are just three TODOs left for stuff that I'm still not positive about.

Patch-set: 2
Reviewer: Gerrit User 13550 <13550@62eb7196-b449-3ce5-99f1-c037f21e1705>