context: Avoid defer in Err()
A common pattern for checking if a context is canceled is to check if the
`Err()` is non-nil. However, this is currently slower than checking if the
channel returned by `Done()` is closed. This is due to the use of defer in
`cancelCtx.Err()`.

Included is a benchmark to demonstrate the improvement:

name                  old time/op  new time/op  delta
CheckCanceled/Err-4   53.5ns ± 2%  20.8ns ± 0%  -61.05%  (p=0.008 n=5+5)
CheckCanceled/Done-4  44.4ns ± 1%  44.5ns ± 0%     ~     (p=0.889 n=5+5)

Change-Id: I2c68700a2b33f8feb3d307ce7c966590a3e960af