Update patch set 35
Patch Set 35:

(3 comments)

> Patch Set 32:
> I assume you just ran "go test" without modifying the "sizes :=" line in testBentleyMcIlroy? Much more sizes than the few present need to be added for exhaustive testing. That said, I did not try testing with coverage analysis, I will try that (with more sizes in the sizes slice) later.

Yes, please do.

> 
> > ... Formally it would be simpler to neglect the movement
> > imitation buffer and implement the actual algorithm from
> > Kim and Kutzner but this probably would not be much faster
> > (if at all for realistic block sizes).
> 
> I do not understand this part of your message. In my code the MIB is used as in the paper, for block rearrangement. To what difference between the paper and my code are you alluding? 

Maybe I'm reading the paper or your code wrong but I
got the impression that Kim&Kutzner's paper use an
"in-data only" MIB while your code introduces an addtional
external one (the one you call "compileTime", which is
clearly superior in regard to speed)

I have to admit, I find it complicated to match the paper's
algorithms given in pseudocode to your code.

> > But as I asked already: Are we sure _this_ algorithm from
> > Kim and Kutzner is the best for merging blocks of _equal_
> > sizes using a MIB? I could imagine a much simpler and faster
> > algorithm.
> 
> It is the best I know of, if you know a better one describe it or reference a paper; and presumably implement it.

Sorry, I do not have a suitable reference at hand.

Stable- Optimal-Block-Merge degenerates to Block-Rotation-Merge
for k ≥ √m/2 (see sec. 3 of the paper). So why not merge blocks
of different sizes? That's what the paper is all about.
The current implementation merges equal size blocks where
n = m therefore k = 1. The papers algorithm is optimal for 
all k>=1 but we do not _need_ an optimal algorithm for a
fixed k as we can _choose_ k (in a winde range). Instead of
merging equal sized blocks (of powers of two) we could split
a block of size N to sort into parts of size sqrt(N) and
N-sqrt(N) (roughly) and merge with a simple Block-Rotation-Merge.

Maybe the algorithm would perform better if k was not == 1
but e.g. 2 or 1.35? Maybe k==1 _is_ optimal but I'd like
to see some supporting evidence for this.

Patch-set: 35